requirements.txt
streamlit==1.42.0
streamlit-authenticator==0.3.3
pandas
reportlab==4.2.2  # for PDF generation
PyPDF2==3.0.1     # optional
python-dotenv
google-api-python-client
google-auth-httplib2
ngoogle-auth-oauthlib
pydrive2   # Optional lighter alternative for Drive
docusign-esign
PyJWT
cryptography